voice_cloning_agent/agent-starter-python-main/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-cloning-agent"
version = "1.0.0"
description = "Real-time telephonic voice cloning agent using LiveKit"
requires-python = ">=3.9"

dependencies = [
    "livekit-agents~=1.2",
    "livekit-plugins-openai",
    "livekit-plugins-assemblyai",
    "python-dotenv",
    "requests",
    "pydub",
    "numpy",
    # Note: scipy and noise-cancellation removed due to 32-bit Python compatibility issues
    # Install 64-bit Python to use these dependencies
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
